# Conflicts:
#	build.gradle
#	grails-app/services/uk/co/wonderlane/wlpos/TillAssignmentService.groovy
#	grails-app/views/cashReporting/tenderMovements.gsp
#	grails-app/views/category/index.gsp
#	grails-app/views/layouts/main.gsp
#	grails-app/views/tag/index.gsp

+++ grails-app/domain/uk/co/wonderlane/wlpos/ProductAttributes.groovy
@@ -60,7 +60,7 @@ class ProductAttributes {
             if (obj.type == ProductAttributeType.NUMERIC && val != null) {
                 if (!val.isNumber()) {
                     return ['productAttribute.numeric.default.not.a.number']
-                } else if (val.toLong() > 999999999) {
+                } else if (val.toDouble() > 999999999) {
                     return ['productAttribute.numeric.default.out.of.range']
                 }
             }
